requirements.txt
numpy
scipy
cma
matplotlib
tensorflow-macos
tensorflow-metal
